fix: thread delete api 변경 및 postMessageAndFiles 수정 및 적용
1. thread delete api 변경함.(관련 file api 삭제함.)
2. postMessageAndFiles 수정 및 적용함. postMessage는 그대로 둠.
